Сдела модульность
